+++ packages/dataviews/src/components/dataform-controls/text.tsx
@@ -2,7 +2,6 @@
  * WordPress dependencies
  */
 import { createElement } from '@wordpress/element';
-import { decodeEntities } from '@wordpress/html-entities';
 
 /**
  * Internal dependencies
@@ -32,7 +31,6 @@ export default function Text< Item >( {
 				validity,
 				prefix: prefix ? createElement( prefix ) : undefined,
 				suffix: suffix ? createElement( suffix ) : undefined,
-				transformValue: decodeEntities,
 			} }
 		/>
 	);